move ereferral class from main to ereferrals variant
